This is a task from frontendmentor.io. It involves the use of HTML and CSS and it was about creating a webpage to display single price grid component.
This is a task from frontendmentor.io. It involves the use of HTML and CSS and it was about creating a webpage to display single price grid component.
However, I was not able to include the how the webpage would respond on a mobile device, as I am not done learning about reponsive design.